refactor: move the lid fine-tuning and label-tab groups into sibling components
LidSection hands its fine-tuning disclosure to LidAdvancedFields, and
LabelTabsSection hands the "tab shape & size" and "engraved text" groups to
LabelTabShapeControls and LabelTabTextStyleControls. Each takes the section
hook's state, handlers and t like LidGripControls does, so the markup moves
verbatim and the hooks are untouched. Each new component has a sibling test
that renders it through the real hook against a seeded store.
